requirements.txt
PyCryptodome
pybluez
